delete infos works

+++ learnTineLearn_backend/learnTineLearn/src/main/java/com/tine/learnTineLearn/InfoRepository.java
@@ -2,6 +2,6 @@ package com.tine.learnTineLearn;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface InfoRepository extends CrudRepository<Info, Integer> {
+public interface InfoRepository extends CrudRepository<Info, Long> {
 }
